Revert "experiment with runtime.dll for relocations"
mingw relocations need to be done outside loader for compat with wine/windows

This reverts commit caae01c59bca47eb803b0159b57b0caea8f286af.

+++ src/programs/windows/win_dynamic/win_dynamic_lib.c
@@ -3,8 +3,8 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-uint32_t lib_var_bss = 0;
-uint32_t lib_var_data = 42;
+uint64_t lib_var_bss = 0;
+uint64_t lib_var_data = 42;
 
 /**
  * Called on stdlib load, unload, etc.
@@ -17,10 +17,10 @@ bool DllMain(
     return true;
 }
 
-uint32_t *get_lib_var_bss() {
+uint64_t *get_lib_var_bss() {
     return &lib_var_bss;
 }
 
-uint32_t *get_lib_var_data() {
+uint64_t *get_lib_var_data() {
     return &lib_var_data;
 }
